remove a bunch of 'Unused variable' warnings.

+++ sph+nln/physics_sph.c
@@ -23,7 +23,7 @@ void SPHFindBbox(SPHbody *bp, int n, float *rmin, float *rmax) {
         VVVV(if LPAREN rmin, > b->pos, RPAREN rmin, = b->pos);
         VVVV(if LPAREN rmax, < b->pos, RPAREN rmax, = b->pos);
         VS(if LPAREN !isfinite LPAREN b->pos,
-           RPAREN RPAREN Error("Bad value for particle %d of %d\n", b - bp, n));
+           RPAREN RPAREN Error("Bad value for particle %d of %d\n", (int)(b - bp), n));
     }
     MPMY_ICombine_Init(&req);
     MPMY_ICombine(rmin, rmin, NDIM, MPMY_FLOAT, MPMY_MIN, req);
